Handling error messages from login / register
Installation of react-toastify for handling error messages from login and register. Changes in utils/helpers for adding handleErrorMessage function. Minor CSS changes for app base style.
