+++ frontend/src/App.js
@@ -1,8 +1,10 @@
 import React, { Component } from "react";
 import { BrowserRouter as Router } from "react-router-dom";
+import { ToastContainer } from "react-toastify";
 
 import BaseRouter from "./routes";
 
+import "react-toastify/dist/ReactToastify.min.css";
 import "./App.css";
 
 class App extends Component {
@@ -12,6 +14,18 @@ class App extends Component {
         <Router>
           <BaseRouter />
         </Router>
+        <ToastContainer
+          position="top-right"
+          autoClose={4000}
+          hideProgressBar
+          newestOnTop
+          closeButton={false}
+          closeOnClick
+          rtl={false}
+          pauseOnVisibilityChange={false}
+          draggable
+          pauseOnHover
+        />
       </div>
     );
   };
