+++ frontend/src/components/Logo.css
@@ -1,6 +1,5 @@
 .logo h1 {
   color: white;
   text-align: center;
-  font-weight: 400;
   margin: 0;
 }
