+++ frontend/src/containers/Home.css
@@ -3,5 +3,5 @@
 }
 
 .home .btn {
-  margin: 0 0.5rem 0.5rem;
+  margin-top: 1rem;
 }
