+++ frontend/src/containers/NoMatch.css
@@ -9,7 +9,6 @@
 .no-match h1 {
   margin: 0 0 0.5rem;
   font-size: 3.5rem;
-  font-weight: 400;
 }
 
 .no-match p {
